LOG-5741: Refactor logstash output functional test
