+++ test/functional/outputs/logstash/suite_test.go
@@ -1,0 +1,13 @@
+package logstash
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestFunctionalLogStashOutput(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "[functional][outputs][logstash] Suite")
+}
